Track account created. Created algo util
